+++ calepin/src/website/site.rs
@@ -193,6 +193,7 @@ impl SiteModel {
             }),
             vars: Default::default(),
             toc_depth: None,
+            toc_floating: None,
         }
     }
 }